requirements.txt
# Flask Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pillow>=10.0.0
scikit-learn>=1.3.0
seaborn>=0.13.0
tqdm>=4.66.0
opencv-python>=4.8.0

# TensorFlow/Keras (updated for compatibility)
tensorflow>=2.16.0
keras>=3.0.0

# High-Accuracy Model Support (optional)
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0